+++ functions/src/actions/user/GetMe.js
@@ -6,10 +6,9 @@ class GetMe extends Exe {
     })
 
     async execute({ userRecord }) {
-        console.log('userRecord', userRecord)
         return {
             data: {
-                userRecord: userRecord,
+                userRecord,
             }
         }
     }