+++ functions/src/actions/user/index.js
@@ -1,7 +1,9 @@
 import GetMe from './GetMe'
 import ChangePassword from './ChangePassword'
+import DeleteMe from './DeleteMe'
 
 export {
     GetMe,
     ChangePassword,
+    DeleteMe,
 }