+++ functions/src/routes/user.js
@@ -1,9 +1,10 @@
 import { MyRouter } from '../helper/router'
-import { GetMe, ChangePassword } from '../actions/user'
+import { GetMe, ChangePassword, DeleteMe } from '../actions/user'
 
 const router = new MyRouter()
 
 router.get("/v1/user/me", GetMe, { authEnabled: true })
+router.delete("/v1/user/me", DeleteMe, { authEnabled: true })
 router.post("/v1/user/password/change", ChangePassword, { authEnabled: true })
 
 export default router