+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - user picture upload
 - user picture delete
 
+- improve custom router
+
 ## [1.0.2] - 2020-10-03
 
 ### Changed